+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Changes the parsed tag name of the default XML namespace from `$foo` to `:foo` to better match other unknown namespaces.
+
 ## 0.1.338 (Feb 6, 2017)
 - Fixes `hugml --version`.
 - Fixes `hugml` name in error message.